created 3 tests for doubling, one for error message one for int one for double

+++ week-09/day-1/src/main/java/com/greenfoxacademy/demo/models/Doubling.java
@@ -5,35 +5,35 @@ import org.springframework.stereotype.Component;
 @Component
 public class Doubling {
 
-  private int received;
-  private int result;
+  private Double received;
+  private Double result;
 
   public Doubling() {
 
   }
 
-  public Doubling(int received) {
+  public Doubling(double received) {
     this.received = received;
     this.result = 2 * received;
   }
 
-  public Integer doubleInput(int received){
+  public Double doubleInput(double received){
     return 2 * received;
   }
 
-  public int getResult() {
+  public Double getResult() {
     return result;
   }
 
-  public void setResult(int result) {
+  public void setResult(double result) {
     this.result = result;
   }
 
-  public int getReceived() {
+  public Double getReceived() {
     return received;
   }
 
-  public void setReceived(int received) {
+  public void setReceived(double received) {
     this.received = received;
   }
